+++ backend/src/main/java/com/lumina_book/backend/dto/response/ProductResponse.java
@@ -23,9 +23,9 @@ public class ProductResponse {
     String author;
     String publisher;
     Double weight;
-    Integer length;
-    Integer width;
-    Integer height;
+    Double length;
+    Double width;
+    Double height;
     Double price;
     Double tax;
     Double discountValue;
